Move ForkPool integration
This can be used by multiple integrations,
not just in Symfony's
